agente/requirements.txt
psycopg2-binary
pynput
psutil
pywin32
https://lookerstudio.google.com/s/hYGufeBhKus